Migrate export and pin methods to host module
These methods could be useful for host use cases in general.

Clean up remaining code and move internal.ui package to root.
