+++ appwidget-viewer/src/main/java/com/google/android/glance/tools/viewer/ui/ViewerPanel.kt
@@ -14,21 +14,8 @@
  * limitations under the License.
  */
 
-package com.google.android.glance.tools.viewer.internal.ui.theme
+package com.google.android.glance.tools.viewer.ui
 
-import androidx.compose.material3.Typography
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.sp
-
-// Set of Material typography styles to start with
-internal val Typography = Typography(
-    bodyLarge = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 16.sp,
-        lineHeight = 24.sp,
-        letterSpacing = 0.5.sp
-    )
-)
+internal enum class ViewerPanel {
+    Resize, Info
+}
