+++ appwidget-viewer/src/main/java/com/google/android/glance/tools/viewer/ui/ViewerResizePanel.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.android.glance.tools.viewer.internal.ui
+package com.google.android.glance.tools.viewer.ui
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
